[Issue #10761] Modify our Task class to abstract away the DB table
